+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add [example script](examples/mustache) to translate Mustache templates.
 * Add support for storing your API Key in a keyring via the `keyring` module.
 * Added a CI check for copyright headers.
+* New languages available: Korean (`'ko'`) and Norwegian (bokmål) (`'nb'`). Add language code constants and tests.
+
+  Note: older library versions also support the new languages, this update only adds new code constants.
 ### Changed
 ### Deprecated
 ### Removed